pyproject.toml
[tool.poetry]
name = "cctracker"
version = "0.1.0"
description = "Tracks Time and tables remaining in the LSFC Commission corner"
authors = ["Alex Hirschfeld"]
license = "GPLv3"
readme = "README.md"
packages = [
    {include = "cctracker/event", from = "components"},
    {include = "cctracker/log", from = "components"},
    {include = "cctracker/server",from = "bases"},
    {include = "cctracker/artist", from = "components"},
    {include = "cctracker/db", from = "components"},
    {include = "development"},
    {include = "cctracker/fs",from = "components"},
    {include = "cctracker/models",from = "components"},
    {include = "cctracker/cache",from = "components"},
]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = ">=0.118.0,<0.119.0"
python-multipart = ">=0.0.20,<0.0.21"
pydantic-settings = ">=2.11.0,<3.0.0"
pydantic-extra-types = ">=2.10.5,<3.0.0"
itsdangerous = ">=2.2.0,<3.0.0"
sqlalchemy = ">=2.0.44,<3.0.0"
pyjwt = { version = ">=2.10.1,<3.0.0", extras = ["crypto"] }
python-keycloak = ">=5.8.1,<6.0.0"
asyncpg = "^0.31.0"
valkey = {extras = ["libvalkey"], version = "^6.1.1"}
minio = "^7.2.20"
tzdata = "^2025.2"
hypercorn = "^0.18.0"

[tool.poetry.group.dev.dependencies]
black = ">=25.9.0,<26.0.0"
isort = ">=6.0.1,<7.0.0"
granian = { version = ">=2.6.0,<3.0.0", extras = ["pname", "uvloop"] }
pytest = "^9.0.2"
aiosqlite = "^0.22.0"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
norecursedirs = ["projects", '*.egg', '.*', '_darcs', 'build', 'CVS', 'dist', 'node_modules', 'venv', '{arch}']
testpaths = ["tests"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["cctracker"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.pyright]
pythonVersion = "3.13"
pythonPlatform = "Linux"
reportMissingImports = true
reportUnusedImport = true
extraPaths = ["bases", "components"]

